requirements.txt
tensorflow_gpu==1.13.0rc1
matplotlib==3.0.2
numpy==1.16.1
scipy==1.2.1
opencv_python==4.0.0.21
Keras==2.2.4
Pillow==5.4.1
selenium==3.141.0
scikit_learn==0.20.2
tensorflow==1.12.0
